fix: audio player and spleeter  when passed non-ascii filename
