+++ electron-frontend/release/app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ktvanywhere",
-  "version": "0.6.0",
+  "version": "0.6.1",
   "description": "A karaoke software",
   "license": "ISC",
   "author": {
