+++ electron-frontend/requirements.txt
@@ -1,4 +1,3 @@
 PyInstaller==5.1
-Quartz==0.0.1.dev0
 setuptools==56.0.0
 spleeter==2.3.0
