+++ electron-frontend/src/components/Song/Song.tsx
@@ -4,7 +4,6 @@ import EditIcon from '@mui/icons-material/Edit';
 import 'react-edit-text/dist/index.css';
 import { Button, Container, Typography } from '@mui/material';
 import LoadingButton from '@mui/lab/LoadingButton';
-import { useSongsStatus } from './SongsStatus.context';
 
 export interface SongProps {
   songId: string;
@@ -54,7 +53,6 @@ interface SongComponentProps {
 const Song = ({ song, setSong }: SongComponentProps) => {
   const [currSong, setCurrSong] = useState(song);
   const [isFetchingLyrics, setIsFetchingLyrics] = useState(false);
-  const { songsStatus, setSongsStatus } = useSongsStatus();
 
   const changeSong = (changedSong: SongProps) => {
     setCurrSong(changedSong);
@@ -99,10 +97,6 @@ const Song = ({ song, setSong }: SongComponentProps) => {
         setIsFetchingLyrics(false);
         return false;
       });
-  };
-
-  const spleeterSeparateVocalsAndMusic = async () => {
-    setSongsStatus([...songsStatus, song.songId]);
   };
 
   return (
@@ -215,15 +209,6 @@ const Song = ({ song, setSong }: SongComponentProps) => {
           onClick={() => getLyrics()}
         >
           fetch lyrics
-        </LoadingButton>
-      </Container>
-      <Container disableGutters>
-        <LoadingButton
-          loading={songsStatus.indexOf(song.songId) !== -1}
-          variant="outlined"
-          onClick={() => spleeterSeparateVocalsAndMusic()}
-        >
-          separate vocals
         </LoadingButton>
       </Container>
     </Container>
